--- server/app.js
@@ -1,32 +1,0 @@
-const express = require('express');
-const admin = require('firebase-admin');
-
-const app = express();
-const serviceAccount = require('./service-account.json');
-
-admin.initializeApp({
-  credential: admin.credential.cert(serviceAccount),
-});
-const db = admin.firestore();
-
-app.get('/users', async (req, res) => {
-  try {
-    const usersCollection = db.collection('users');
-    const snapshot = await usersCollection.get();
-
-    const users = [];
-    snapshot.forEach((doc) => {
-      users.push({ id: doc.id, ...doc.data() });
-    });
-
-    res.status(200).json(users);
-  } catch (error) {
-    console.error('Error fetching users:', error);
-    res.status(500).send('Error fetching users');
-  }
-});
-
-const PORT = process.env.PORT || 5000;
-app.listen(PORT, () => {
-  console.log(`Server is running on port ${PORT}`);
-});
