+++ server/src/config/firebaseAdmin.js
@@ -1,0 +1,9 @@
+const admin = require('firebase-admin');
+const serviceAccount = require('../../service-account.json');
+
+admin.initializeApp({
+  credential: admin.credential.cert(serviceAccount),
+});
+const db = admin.firestore();
+
+module.exports = db;
