+++ server/src/routes/usersRoutes.js
@@ -1,0 +1,7 @@
+const express = require('express');
+const router = express.Router();
+const usersController = require('../controllers/usersController');
+
+router.get('/', usersController.getUsers);
+
+module.exports = router;
